fix(tests): break heartbeat-filter -> node:* import chain
The browser-import boundary test
(src/auto-reply/heartbeat-filter.browser-import.test.ts) failed
because the heartbeat-filter dependency graph reaches node:fs,
node:path, node:os, node:async_hooks, node:crypto via:

  heartbeat-filter.ts
    -> heartbeat.ts
    -> tokens.ts
    -> infra/diagnostic-trace-context.ts (node:async_hooks, node:crypto)
    -> auto-reply/continuation/targeting.ts
       -> infra/session-delivery-queue-storage.ts
       -> config/paths.ts (node:fs/os/path)
       -> channels/bundled-channel-catalog-read.ts (node:fs/path)

tokens.ts only needs pure helpers from each leak: a regex/parser
and a few small normalizers. Extracting those into pure sibling
files (-pure.ts) and pointing tokens.ts at the pure imports breaks
the chain without changing any runtime behavior.

- New src/auto-reply/continuation/targeting-pure.ts holds the
  fanout-mode constant, the targeting type aliases, and the three
  normalize/has helpers. targeting.ts re-exports for back-compat
  with existing consumers.
- New src/infra/diagnostic-trace-context-pure.ts holds the
  traceparent regex/parsers and validators. diagnostic-trace-context.ts
  re-exports the same names so existing consumers do not need to move.
- tokens.ts now imports directly from the two pure files.

No code logic changes; only file boundaries shifted to keep
node-only imports out of the browser-bundle reachable graph.
